Renamed deleteUser to deleteProfile and moved it to ProfilesController
